+++ src/design-system/assets-table/assets-table.stories.tsx
@@ -49,11 +49,27 @@ interface AssetInfoProps {
 const SimpleAssetInfo = ({ id }: Readonly<AssetInfoProps>): JSX.Element => (
   <AssetsTable id={id}>
     <TokenProfile
-      imageSrc={cardanoImage}
+      image={cardanoImage}
       name="Token name"
       description="Subtitle"
     />
-    <TokenAmount amount="23,584.48" fiatPrice="24,568.54 USD" />
+    <TokenAmount amount="23,584.48" details="24,568.54 USD" />
+  </AssetsTable>
+);
+
+const PendingAssetInfo = ({ id }: Readonly<AssetInfoProps>): JSX.Element => (
+  <AssetsTable id={id}>
+    <TokenProfile
+      image={cardanoImage}
+      name="Token with pending amount"
+      description="description"
+    />
+    <TokenAmount
+      amount="21,584.48"
+      details="+ 5 pending"
+      detailsColor="success"
+      detailsTooltip="Pending coins represent tokens that the wallet recognizes but are not yet finalized by the network. These coins are currently unspendable until confirmed."
+    />
   </AssetsTable>
 );
 
@@ -63,7 +79,7 @@ const DetailedAssetInfo = ({
 }: Readonly<AssetInfoProps>): JSX.Element => (
   <AssetsTable id={id}>
     <TokenProfile
-      imageSrc={cardanoImage}
+      image={cardanoImage}
       name="Token name"
       description="Subtitle"
     />
@@ -72,7 +88,7 @@ const DetailedAssetInfo = ({
       priceChange="+3.21"
       priceTrend={priceTrend}
     />
-    <TokenAmount amount="23,584.48" fiatPrice="24,568.54 USD" />
+    <TokenAmount amount="23,584.48" details="24,568.54 USD" />
   </AssetsTable>
 );
 
@@ -124,6 +140,11 @@ export const Overview = (): JSX.Element => (
           <Variants.Row>
             <Variants.Cell>
               <DetailedAssetInfo />
+            </Variants.Cell>
+          </Variants.Row>
+          <Variants.Row>
+            <Variants.Cell>
+              <PendingAssetInfo />
             </Variants.Cell>
           </Variants.Row>
         </Variants.Table>
@@ -179,7 +200,7 @@ export const Controls: Controls = ({
         <Flex alignItems="center" flexDirection="column">
           <AssetsTable>
             <TokenProfile
-              imageSrc={cardanoImage}
+              image={cardanoImage}
               name={tokenName}
               description={tokenSubtitle}
             />
@@ -188,7 +209,7 @@ export const Controls: Controls = ({
               priceChange={priceChange}
               priceTrend={priceTrend}
             />
-            <TokenAmount amount={tokenAmount} fiatPrice={fiatPrice} />
+            <TokenAmount amount={tokenAmount} details={fiatPrice} />
           </AssetsTable>
         </Flex>
       </Section>