+++ src/design-system/profile-dropdown/profile-dropdown-wallet-option.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import { action } from '@storybook/addon-actions';
 import type { Meta } from '@storybook/react';
 
 import { page, Variants, Section, UIStateTable } from '../decorators';
@@ -65,15 +66,14 @@ export const Overview = (): JSX.Element => (
             <Variants.Row>
               <Variants.Cell>
                 <WalletOption
-                  title="Alice's wallet"
+                  title="WALLET WITH REALLY LONG NAME THAT WILL BE TRUNCATED"
                   subtitle="Account #0"
                   type="hot"
                 />
               </Variants.Cell>
               <Variants.Cell>
                 <WalletOption
-                  title="Alice's wallet"
-                  subtitle="Account #0"
+                  title="WALLET WITH REALLY LONG NAME THAT WILL BE TRUNCATED"
                   type="cold"
                 />
               </Variants.Cell>
@@ -82,6 +82,44 @@ export const Overview = (): JSX.Element => (
                   title="Shared Wallet"
                   subtitle="Lace Team"
                   type="shared"
+                />
+              </Variants.Cell>
+            </Variants.Row>
+          </Variants.Table>
+        </Flex>
+      </Section>
+
+      <Divider my="$64" />
+
+      <Section title="Example with actions">
+        <Flex flexDirection="column" alignItems="center" w="$fill">
+          <Variants.Table
+            headers={['Hot wallet', 'Cold wallet', 'Shared wallet']}
+          >
+            <Variants.Row>
+              <Variants.Cell>
+                <WalletOption
+                  title="Alice's wallet"
+                  subtitle="Account #0"
+                  type="hot"
+                  onOpenAccountsMenu={action('onOpenAccountsMenu')}
+                  onOpenEditWallet={action('onOpenEditWallet')}
+                />
+              </Variants.Cell>
+              <Variants.Cell>
+                <WalletOption
+                  title="WALLET WITH REALLY LONG NAME THAT WILL BE TRUNCATED"
+                  type="cold"
+                  onOpenAccountsMenu={action('onOpenAccountsMenu')}
+                  onOpenEditWallet={action('onOpenEditWallet')}
+                />
+              </Variants.Cell>
+              <Variants.Cell>
+                <WalletOption
+                  title="Shared Wallet"
+                  subtitle="Lace Team"
+                  type="shared"
+                  onOpenEditWallet={action('onOpenEditWallet')}
                 />
               </Variants.Cell>
             </Variants.Row>