+++ src/icons/ExclamationCircleComponent.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import type { SVGProps } from 'react';
+const SvgExclamationCirclecomponent = (props: SVGProps<SVGSVGElement>) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={113}
+    height={113}
+    fill="none"
+    {...props}
+  >
+    <path
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth={3}
+      d="M56.5 37.455v18.667m0 18.667h.047M98.5 56.122c0 23.196-18.804 42-42 42s-42-18.804-42-42 18.804-42 42-42 42 18.804 42 42"
+    />
+  </svg>
+);
+export default SvgExclamationCirclecomponent;